Chess - first commit!
Chess is going along very smoothly! The game is not playable yet, but when you click on a piece, its legal moves will be output to the terminal.
PS: if you comment out the pawn creation in ChessBoard.py (lines 20-23), you can see the legal moves of each piece much better.

TODO:
castling - If the king has not yet moved, it is allowed to travel two spaces towards either rook that also has not yet moved, with the rook moving to the other side of the king. For example, long castles (denoted as 0-0-0) starts with the rook on a1 and a king on e1, and ends with the king on c1 and the rook on d1. Short castles (denoted as 0-0) starts with the king on e1 and the rook on h1, and ends with the king in g1 and the rook on f1.

en passant - If a white pawn moved from d2 to d3, a black pawn on c4 is allowed to capture the white pawn now on d3. If the same white pawn moves from d2 to d4, then the black pawn on c4 is still allowed to capture the white pawn now on d4, provided that the black pawn captures the white pawn on the very next move.

check - If the king is being attacked, that is, a piece of opposite color is threatening to capture the king on the next move, the king is said to be in check. This takes priority over any legal move of the player in check, and the player must respond to the check by (a) blocking the path between the attacking piece and the king, (b) capturing the piece giving the check, or (c) moving the king to a safe square.

Known bugs:
none