Chess - gameplay finished! needs polishing
The game is now fully playable! Castling functions as intended, and so does en passant. However, when in check, the king is not forced to move. Instead, if you capture the opponent's king, you win.
TODO:
i. add PAUL bot - You will be able to play 1-on-1 locally (current version limited to this game mode) or play against PAUL himself. PAUL will make a random legal move for each of his turns, making his playstyle very tricky and unorthodox.
ii. polish game & visuals - with the PAUL bot will come an icon at the top showing his face, with sprites for idling, thinking, winning, and losing. A PAUse menu may be implemented as well.
iii. alternate piece sets - Currently, the only piece set is letters that reference the pieces. At some point I will draw sprites for my own piece set.

Possible modifications/additions:
i. check - As stated before, the king is not forced to move when in check. This may change depending on how everyone would prefer the game to function.
ii. move notation - Chess has a common notation system used to list the moves of a game. I'd like to add this in a text field to the right of the board.

Known bugs:
none