Chess - check, 1, 2, check...
We've got sound effects in chess, folks! The rest of it still plays great, and now, sounds great!

Additions:
i. background music - I have some classy jazz music paying in the background. This is a placeholder until Matthew and I get around to writing the soundtrack.
ii. sound effects - Sound effects for moving a piece, capturing another piece, and castling have been added. Gone are the days of boring, silent PAUL chess!
iii. pawn promotion - Previously, when a pawn reached the back rank, nothing would happen. In chess, when a pawn reaches its opponent's back rank, it may promote into a queen. Promotion to a rook, bishop, or knight is also allowed. For now, the game automatically promotes a pawn to a queen when it reaches the appropriate back rank.

TODO:
i. text on screen - There are a few text objects I'd like to add. First off, the terminal sees the "White wins!" and "Black wins!" text. Secondly, PAUL's lines will need to be shown on the game window as well. Lastly, the "play again" text for the menu will be added.
ii. PAUL personality - right now, all I have for PAUL is his icon with facial expressions. I'm going to add some lines for him to say, with separate lines for capturing a piece, losing a piece, winning, and losing,
iii. move notation - I may have previously mentioned a desire to add this, but having the notation of the moves played on the right-hand side is a planned addition. This will allow the player (and PAUL) to see the full game's moves.
iv. "play again" menu - Once the game is over, a "play again" menu will appear, either over the board or over the notation on the right of the window.

Known bugs:
incorrect sfx - I had a few issues with moving pawns and playing the 'capture' sound effect instead of the 'move' sound effect. I changed the code to try and fix it, so this may or may not be resolved.