Platformer - first commit!
Progress for the platformer is officially underway!

files:
Main.py - This function serves as a placeholder menu for the platformer. This will eventually fade away into the main menu in PAUL.
Tile.py - This class creates Tile objects, which construct each individual level.
Levels.py - This class creates Level objects, generating tiles from a 15x20 array in the file itself. These arrays will later come from a levels.json file.
Player.py - This class creates a Player object, which is controlled by the player.
Game.py - This function is the blood of the platformer, tying everything together.

additions:
i. player: PAUL has been born into the game world as a playable character in this platformer. The 'left' and 'right' arrow keys are for moving sideways, and the player can jump with 'up' or 'space'.
ii. tiles: each individual tile has its own sprite. The code for tiles goes as follows:
    0 - empty: This is an empty, non-colliding tile. Pretty straightforward.
    1 - dirt: This is a dirt-filled, colliding tile. Dirt serves the purpose of a simple ground block.
    2 - lava: This is a molten, colliding tile. If the player collides with lava, the player is sent back to the beginning of the level, and the player's death counter is incremented.
    3 - goal: This is a successful, colliding tile. If the player collides with the goal, the player completes the level.
    4 - coin: This is a shiny, non-colliding tile. If the player collides with a coin, then the coin disappears and the player's coin counter is incremented.
    5 - sign: This is a legible, non-colliding tile. If the player stands over a sign and presses '[I haven't decided what key yet]', the player can read the contents of the sign.
iii. levels: Levels in this platformer are generated from a 15x20 array of numbers. These numbers will generate tiles based on the key in (ii.).
iii. game loop: The game loop ties all of these classes together to create a cohesive platforming experience. The game_loop(level) function takes one parameter, dictating what level the player starts on.

TODO:
i. sign work - The key to read the sign has yet to be decided, as well as how to store the contents of the sign for the player to read.
ii. level progression - Currently, when you reach the goal post, the terminal sees "You win!", as well as a display of coins collected and deaths counted. This is a placeholder, but theoretically game_loop(current_level.level_num+1) will play the next level.
iii. timer (?) - A timer could possibly be used for the leaderboard. The leaderboard function for the platformer could keep track of coins collected, shortest time completion, or fewest deaths.
iv. levels.json - Currently, levels are stored in Levels.py. This new .json file will be used in the future to store different level layouts.