Platformer - collision mayhem
The platformer's collision is now much smoother to play!

additions:
i. sign work - The signs in the game are now functional! Press the 'c' key to read the contents of the sign.
ii. levels - The levels now have backgrounds corresponding to each of the other PAUL games (with the current exception of mad libs). The levels also automatically progress once you reach the goal.
iii. better collision - The PAUL sprite only took up a 34x38 rectangle, as opposed to 30x30 for the rest of the tiles. Now, PAUL no longer gets hung up on corners.
iv. optimization - The code is now more efficient in several areas.

removal:
levels.json - This file turned out to be redundant, as it was easier to keep the level formats in Levels.py, as opposed to having to access another file needlessly.

TODO:
i. timer (?) - A timer could possibly be used for the leaderboard. The leaderboard function for the platformer could keep track of coins collected, shortest time completion, or fewest deaths.
ii. more levels - Currently, the game has 2 fully finished levels, with the exception of the background for "Level 2: Checkers".