Scoreboard
Isn't connected to main menu right now and is not completely done. However it is a skeleton so we just need to update the JSON to the proper structure and it should work, need to have to list more scores in the JSON but we can deal with that later. Also we will have to work with how we want to access it just because it needs to stay above the games file wise to access them. :/